tools/doc_converter/pyproject.toml
[project]
name = "doc-converter"
version = "0.1.0"
description = "Document → Markdown local OSS pipeline + DoXA adapter"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    # 2026-04-22 Python 3.12.12 + uv 환경에서 함께 설치 확인.
    # 외부 AI SDK (anthropic/openai/google-genai) 끌어오는 marker-pdf 제외.
    "mineru[core]>=3.1,<4",
    "docling>=2.90,<3",
    "markitdown[pdf,pptx,docx]>=0.1,<1",
    "easyocr>=1.7",
    "pymupdf>=1.27",
    "pandas>=2.2",
    "openpyxl>=3.1",
    "tabulate>=0.10",
    "requests>=2.32",
    "urllib3>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
doxa = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
include = [
    "*.py",
    "*.bat",
    "*.md",
    "requirements.txt",
]
